+++ app/Http/Controllers/Cms/Hotels/CmsHotelsController.php
@@ -27,18 +27,22 @@ class CmsHotelsController extends BaseCmsHotelsController
     public function store(StoreHotelRequest $request): RedirectResponse
     {
         $validatedData = $request->validated();
-        $this->getHotelService()->store($validatedData);
+        $this->getHotelService()->store($validatedData, $request->user());
 
         return redirect()->route(CmsRoutes::CMS_HOTELS_INDEX);
     }
 
     public function edit(Hotel $hotel): View
     {
+        $this->authorize('view', $hotel);
+
         return view('cms.hotels.edit', compact(['hotel']));
     }
 
     public function update(UpdateHotelRequest $request, Hotel $hotel): RedirectResponse
     {
+        $this->authorize('update', $hotel);
+
         $validatedData = $request->validated();
         $this->getHotelService()->update($validatedData, $hotel);
 
@@ -47,6 +51,8 @@ class CmsHotelsController extends BaseCmsHotelsController
 
     public function destroy(Hotel $hotel): RedirectResponse
     {
+        $this->authorize('delete', $hotel);
+
         $this->getHotelService()->delete($hotel);
 
         return redirect()->route(CmsRoutes::CMS_HOTELS_INDEX);
